+++ app/routes/dashboard/home.js
@@ -1,6 +1,5 @@
 import Route from '@ember/routing/route';
 import { service } from '@ember/service';
-import { action } from '@ember/object';
 import { isBlank } from '@ember/utils';
 
 import RSVP from 'rsvp';
@@ -34,10 +33,5 @@ export default class DashboardHomeRoute extends Route {
         backgroundReload: true,
       }),
     });
-  }
-
-  @action
-  loading() {
-    return true;
   }
 }